requirements.txt
# AutoGen for agent system
pyautogen==0.2.0

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# YOLO
ultralytics==8.0.200

# LLM APIs
openai==1.3.0
anthropic==0.64.0

# Google Cloud Vision
google-cloud-vision==3.4.5

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0

# Image processing
scikit-image==0.22.0
matplotlib==3.8.0

# Data handling
pandas==2.1.3
pydantic==2.4.2

# MongoDB
pymongo==4.6.0

# PDF Processing
PyMuPDF==1.23.8

# Utilities
requests==2.31.0
aiohttp==3.9.0